requirements.txt
langchain
ollama
langchain-community
arxiv
chromadb
sentence-transformers
pdfplumber
streamlit
python-dotenv
